requirements.txt
anyio==4.0.0
APScheduler==3.6.3
asgiref==3.7.2
cachetools==4.2.2
certifi==2023.7.22
Django==4.2.6
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
mypy==1.6.1
mypy-extensions==1.0.0
python-telegram-bot==13.7
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzlocal==5.2
